重构项目结构，改进模块化设计，添加命令行参数支持
