+++ server/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = "server"
-version = "0.5.1"
+version = "0.6.0"
 edition = "2021"
-authors = ["WisteFinch", "FloatingLight233"]
+authors = ["WisteFinch", "FloatingLight233", "djkcyl"]
 license = "MIT"
 repository = "https://github.com/WisteFinch/Helldivers2CallForStratagemsOnPhone"
 
@@ -13,12 +13,14 @@ rdev = "0.5.3"
 tokio = { version = "1", features = ["full"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+toml = "0.8.10"
 local_ipaddress = "0.1.3"
 colored = "2.1.0"
 rand = "0.8.5"
 fast_qr = "0.12.5"
 rust-i18n = "3.0.1"
 sys-locale = "0.3.1"
+clap = { version = "4.5.4", features = ["derive"] }
 
 [profile.release]
 lto = true
