+++ server/src/data/auth.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Clone)]
+pub struct Auth {
+    pub sid: String,
+    pub time: u64,
+}
