+++ server/src/data/mod.rs
@@ -1,0 +1,11 @@
+// 定义子模块
+pub mod config;
+pub mod input;
+pub mod operation;
+pub mod auth;
+
+// 从子模块重新导出公共API
+pub use config::{AppConfig, AuthConfig, Config, InputConfig, ServerConfig};
+pub use input::{InputData, InputType, KeyType, Step};
+pub use operation::{Operation, Status};
+pub use auth::Auth;
