+++ server/src/locales/zh-CN.yml
@@ -16,6 +16,8 @@ n_scan_qr_code: '使用手机端扫描二维码可以快速获取连接配置'
 warn_conf_load_failed: '读取配置失败，使用默认配置！'
 warn_conf_network_temp: '使用临时网络配置'
 info_conf_loaded: '已加载配置文件'
+info_conf_migrated: '已成功将旧配置文件迁移到新格式'
+err_conf_serialize_failed: '序列化配置文件失败'
 warn_reject_sync: '拒绝同步请求！'
 n_auth_2: '，sid：'
 n_auth_1: '认证请求自：'
@@ -27,6 +29,8 @@ info_conf_saved: '配置已保存'
 err_sync_failed: '同步配置失败！'
 err_conf_save_failed: '保存配置失败！'
 err_auth_save_failed: '认证信息保存失败！'
+err_old_auth_remove_failed: '删除旧认证文件失败！'
+err_old_conf_remove_failed: '删除旧配置文件失败！'
 err_parse_step_failed: '解析步骤失败！'
 err_parse_input_failed: '解析输入类型失败！'
 n_free_input: '自由输入：'
@@ -35,6 +39,8 @@ n_warn: '[警告] '
 n_info: '[信息] '
 n_debug: '[调试] '
 debug_mode: '已启用调试模式'
+auth_disabled: '已禁用SID认证'
+info_auth_disabled: '认证已禁用，客户端自动通过认证'
 d_remove_redundant: '多余请求已剔除：'
 d_specific_ip: '已指定ip地址：'
 n_conf_title: '当前配置如下'
@@ -48,3 +54,19 @@ n_conf_input_right: '右：'
 n_conf_type: '输入类型'
 n_conf_type_open: '打开战略配备：'
 n_admin: '如游戏内无反应，可尝试使用管理员身份启动'
+ask_migrate_title: '发现旧版配置文件，是否迁移到新格式？'
+ask_migrate: '是否迁移配置到新格式？默认是 (Y/N)：'
+migrate_cancelled: '迁移已取消，继续使用旧配置'
+migrate_required: '必须迁移到新配置格式才能继续使用，程序将退出'
+old_config_info: '旧配置信息'
+auth_records_info: '认证记录数量'
+config_port: '端口'
+config_ip: 'IP地址'
+config_ip_auto: '自动获取'
+config_delay: '输入延迟'
+config_open: '打开键'
+config_open_type: '打开方式'
+config_key_up: '上键'
+config_key_down: '下键'
+config_key_left: '左键'
+config_key_right: '右键'
